Move normalize_answer from search_env.py to tools.py
